cdn/load_service/requirements.txt
fastapi==0.95.0
celery==5.2.7
uvicorn==0.21.1
flower==1.2.0
minio==7.1.14
boto3==1.26.104
orjson==3.8.9
pydantic==1.10.7
redis==4.5.4
gunicorn==20.1.0
httpx==0.24.0
backoff==2.2.1
pytest==7.3.1
pytest-mock==3.10.0
python-dotenv==0.19.2
python-multipart==0.0.6
aiofiles==23.1.0